+++ test/integration/aem_spec.rb
@@ -35,4 +35,22 @@ describe 'Aem' do
 
   end
 
+  describe 'test get_aem_health_check_wait_until_ok' do
+
+    it 'should try once and contain readyness indicator' do
+      aem = @aem.aem()
+      result = aem.get_aem_health_check_wait_until_ok('shallow', false, {
+        _retries: {
+          max_tries: 60,
+          base_sleep_seconds: 2,
+          max_sleep_seconds: 2
+        }})
+      expect(result.message).to eq('AEM health check retrieved')
+      expect(result.data.length).to eq(1)
+      expect(result.data[0]['name']).to eq('Smoke Health Check')
+      expect(result.data[0]['status']).to eq('OK')
+    end
+
+  end
+
 end