+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.0.17
+* Allow get_aem_health_check method parameters tags and combine_tags_or optional
+* Allow get_aem_health_check_wait_until_ok method parameters tags and combine_tags_or optional
+
 ### 1.0.16
 * Add SSL support for postAgent operation
 * Add AEM Health Check Servlet config support